+++ lib/src/providers/providers.dart
@@ -4,6 +4,6 @@
 
 export 'llm_provider_interface.dart';
 export 'echo_provider.dart';
-export 'firebase_vertex_provider.dart';
+export 'vertex_provider.dart';
 export 'gemini_provider.dart';
 export 'ollama_provider.dart';